+++ database/migrations/2016_09_12_000000_rename_like_counter_table.php
@@ -10,13 +10,12 @@
  */
 
 use Illuminate\Support\Facades\Schema;
-use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
 /**
- * Class CreateEntityTable.
+ * Class RenameLikeCounterTable.
  */
-class CreateEntityTable extends Migration
+class RenameLikeCounterTable extends Migration
 {
     /**
      * Run the migrations.
@@ -25,11 +24,7 @@ class CreateEntityTable extends Migration
      */
     public function up()
     {
-        Schema::create('entity', function (Blueprint $table) {
-            $table->increments('id');
-            $table->string('name');
-            $table->timestamps();
-        });
+        Schema::rename('like_counter', 'likers_counter');
     }
 
     /**
@@ -39,6 +34,6 @@ class CreateEntityTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('entity');
+        Schema::rename('likers_counter', 'like_counter');
     }
 }