+++ src/Events/ModelWasUndisliked.php
@@ -39,7 +39,6 @@ class ModelWasUndisliked
      *
      * @param \Cog\Likeable\Contracts\HasLikes $model
      * @param int $likerId
-     * @return void
      */
     public function __construct(HasLikesContract $model, $likerId)
     {