+++ src/Observers/LikeObserver.php
@@ -12,10 +12,10 @@
 namespace Cog\Likeable\Observers;
 
 use Cog\Likeable\Enums\LikeType;
-use Cog\Likeable\Events\ModelWasDisliked;
 use Cog\Likeable\Events\ModelWasLiked;
-use Cog\Likeable\Events\ModelWasUndisliked;
 use Cog\Likeable\Events\ModelWasUnliked;
+use Cog\Likeable\Events\ModelWasDisliked;
+use Cog\Likeable\Events\ModelWasUndisliked;
 use Cog\Likeable\Contracts\Like as LikeContract;
 use Cog\Likeable\Contracts\LikeableService as LikeableServiceContract;
 
@@ -36,10 +36,10 @@ class LikeObserver
     {
         if ($like->type_id == LikeType::LIKE) {
             event(new ModelWasLiked($like->likeable, $like->user_id));
-            app(LikeableServiceContract::class)->incrementLikesCount($like->likeable, $like->type_id);
+            app(LikeableServiceContract::class)->incrementLikersCount($like->likeable, $like->type_id);
         } else {
             event(new ModelWasDisliked($like->likeable, $like->user_id));
-            app(LikeableServiceContract::class)->incrementDislikesCount($like->likeable, $like->type_id);
+            app(LikeableServiceContract::class)->incrementDislikersCount($like->likeable, $like->type_id);
         }
     }
 
@@ -53,10 +53,10 @@ class LikeObserver
     {
         if ($like->type_id == LikeType::LIKE) {
             event(new ModelWasUnliked($like->likeable, $like->user_id));
-            app(LikeableServiceContract::class)->decrementLikesCount($like->likeable, $like->type_id);
+            app(LikeableServiceContract::class)->decrementLikersCount($like->likeable, $like->type_id);
         } else {
             event(new ModelWasUndisliked($like->likeable, $like->user_id));
-            app(LikeableServiceContract::class)->decrementDislikesCount($like->likeable, $like->type_id);
+            app(LikeableServiceContract::class)->decrementDislikersCount($like->likeable, $like->type_id);
         }
     }
 }