+++ src/Providers/LikeableServiceProvider.php
@@ -11,15 +11,15 @@
 
 namespace Cog\Likeable\Providers;
 
-use Cog\Likeable\Console\LikeableRecountCommand;
-use Cog\Likeable\Contracts\Like as LikeContract;
-use Cog\Likeable\Contracts\LikeableService as LikeableServiceContract;
-use Cog\Likeable\Contracts\LikeCounter as LikeCounterContract;
 use Cog\Likeable\Models\Like;
-use Cog\Likeable\Models\LikeCounter;
+use Cog\Likeable\Models\LikersCounter;
+use Illuminate\Support\ServiceProvider;
 use Cog\Likeable\Observers\LikeObserver;
 use Cog\Likeable\Services\LikeableService;
-use Illuminate\Support\ServiceProvider;
+use Cog\Likeable\Console\LikeableRecountCommand;
+use Cog\Likeable\Contracts\Like as LikeContract;
+use Cog\Likeable\Contracts\LikersCounter as LikersCounterContract;
+use Cog\Likeable\Contracts\LikeableService as LikeableServiceContract;
 
 /**
  * Class LikeableServiceProvider.
@@ -54,7 +54,7 @@ class LikeableServiceProvider extends ServiceProvider
     public function register()
     {
         $this->app->bind(LikeContract::class, Like::class);
-        $this->app->bind(LikeCounterContract::class, LikeCounter::class);
+        $this->app->bind(LikersCounterContract::class, LikersCounter::class);
         $this->app->singleton(LikeableServiceContract::class, LikeableService::class);
     }