+++ tests/migrations/2016_09_02_173301_create_entity_with_morph_map_table.php
@@ -9,14 +9,20 @@
  * file that was distributed with this source code.
  */
 
-use Illuminate\Database\Migrations\Migration;
+use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
 
 /**
  * Class CreateEntityWithMorphMapTable.
  */
 class CreateEntityWithMorphMapTable extends Migration
 {
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
     public function up()
     {
         Schema::create('entity_with_morph_map', function (Blueprint $table) {
@@ -26,8 +32,13 @@ class CreateEntityWithMorphMapTable extends Migration
         });
     }
 
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
     public function down()
     {
-        Schema::drop('entity_with_morph_map');
+        Schema::dropIfExists('entity_with_morph_map');
     }
 }